+++ Cosette/Engine/Ai/Ordering/HistoryHeuristic.cs
@@ -11,34 +11,34 @@ namespace Cosette.Engine.Ai.Ordering
         static HistoryHeuristic()
         {
             _historyMoves = new uint[2][][];
-            _historyMoves[Color.White] = new uint[64][];
-            _historyMoves[Color.Black] = new uint[64][];
+            _historyMoves[Color.White] = new uint[6][];
+            _historyMoves[Color.Black] = new uint[6][];
 
-            for (var from = 0; from < 64; from++)
+            for (var piece = 0; piece < 6; piece++)
             {
-                _historyMoves[Color.White][from] = new uint[64];
-                _historyMoves[Color.Black][from] = new uint[64];
+                _historyMoves[Color.White][piece] = new uint[64];
+                _historyMoves[Color.Black][piece] = new uint[64];
             }
 
             _max = MoveOrderingConstants.HistoryHeuristicMaxScore;
         }
 
-        public static void AddHistoryMove(int color, int from, int to, int depth)
+        public static void AddHistoryMove(int color, int piece, int to, int depth)
         {
-            var newValue = _historyMoves[color][from][to] + (uint)(depth * depth);
+            var newValue = _historyMoves[color][piece][to] + (uint)(depth * depth);
 
             _max = Math.Max(_max, newValue);
-            _historyMoves[color][from][to] = newValue;
+            _historyMoves[color][piece][to] = newValue;
         }
 
-        public static short GetMoveValue(int color, int from, int to)
+        public static short GetMoveValue(int color, int piece, int to)
         {
-            return (short)Math.Ceiling(_historyMoves[color][from][to] / ((float)_max / MoveOrderingConstants.HistoryHeuristicMaxScore));
+            return (short)Math.Ceiling(_historyMoves[color][piece][to] / ((float)_max / MoveOrderingConstants.HistoryHeuristicMaxScore));
         }
 
-        public static uint GetRawMoveValue(int color, int from, int to)
+        public static uint GetRawMoveValue(int color, int piece, int to)
         {
-            return _historyMoves[color][from][to];
+            return _historyMoves[color][piece][to];
         }
 
         public static uint GetMaxValue()
@@ -50,11 +50,11 @@ namespace Cosette.Engine.Ai.Ordering
         {
             for (var color = 0; color < 2; color++)
             {
-                for (var from = 0; from < 64; from++)
+                for (var piece = 0; piece < 6; piece++)
                 {
                     for (var to = 0; to < 64; to++)
                     {
-                        _historyMoves[color][from][to] /= 2;
+                        _historyMoves[color][piece][to] /= 2;
                     }
                 }
             }
@@ -67,9 +67,9 @@ namespace Cosette.Engine.Ai.Ordering
         {
             for (var color = 0; color < 2; color++)
             {
-                for (var from = 0; from < 64; from++)
+                for (var piece = 0; piece < 6; piece++)
                 {
-                    Array.Clear(_historyMoves[color][from], 0, 64);
+                    Array.Clear(_historyMoves[color][piece], 0, 64);
                 }
             }