+++ CHANGELOG.md
@@ -18,6 +18,7 @@
  - Added detection of open file next to the king in the evaluation
  - Fixed memory usage issues during search
  - Added history heuristic veto in LMR
+ - Improve history table structure
 
 # Version 4.0 (Komekko), 22.01.2021
  - Added Texel project